C++ 17 compatible (#34)
* C++ 17 compatible

* CI

* C++ 17 compatible

+++ application/components/SIG/SIG_Hub.h
@@ -16,11 +16,11 @@ public:
     inline SIG_Hub()
     {}
 
-    void toFld(UINT32 id, INT32 state, INT32 speed) const;
-    void toGui(UINT32 id, INT32 state, INT32 speed) const;
+    void toFld(size_t id, INT32 state, INT32 speed) const;
+    void toGui(size_t id, INT32 state, INT32 speed) const;
 
-    void fromDsp(UINT32 pos, const FldState& tele) const;
-    void fromDsp(UINT32 pos, const GuiCmd&   tele) const;
+    void fromDsp(size_t pos, const FldState& tele) const;
+    void fromDsp(size_t pos, const GuiCmd&   tele) const;
 
     IL_INSTANCE_DEC(SIG_Hub)
