+++ application/components/TSW/TSW_Hub.h
@@ -16,11 +16,11 @@ public:
     inline TSW_Hub()
     {}
 
-    void toFld(UINT32 id, INT32 cmd) const;
-    void toGui(UINT32 id, INT32 state) const;
+    void toFld(size_t id, INT32 cmd) const;
+    void toGui(size_t id, INT32 state) const;
 
-    void fromDsp(UINT32 pos, const FldState& tele) const;
-    void fromDsp(UINT32 pos, const GuiCmd&   tele) const;
+    void fromDsp(size_t pos, const FldState& tele) const;
+    void fromDsp(size_t pos, const GuiCmd&   tele) const;
 
     IL_INSTANCE_DEC(TSW_Hub)
