+++ specification/ifs/I_SIG_Hub.h
@@ -11,11 +11,11 @@
 class I_SIG_Hub
 {
 public:
-    virtual void toFld(UINT32 id, INT32 state, INT32 speed) const = 0;
-    virtual void toGui(UINT32 id, INT32 state, INT32 speed) const = 0;
+    virtual void toFld(size_t id, INT32 state, INT32 speed) const = 0;
+    virtual void toGui(size_t id, INT32 state, INT32 speed) const = 0;
 
-    virtual void fromDsp(UINT32 pos, const FldState& tele) const = 0;
-    virtual void fromDsp(UINT32 pos, const GuiCmd&   tele) const = 0;
+    virtual void fromDsp(size_t pos, const FldState& tele) const = 0;
+    virtual void fromDsp(size_t pos, const GuiCmd&   tele) const = 0;
 };
 
 #endif // H_
