+++ specification/ifs/I_SIG_Provider.h
@@ -13,8 +13,8 @@
 class I_SIG_Provider
 {
 public:
-    virtual bool has(UINT32 pos) const = 0;
-    virtual I_SIG& at(UINT32 pos) = 0;
+    virtual bool has(size_t pos) const = 0;
+    virtual I_SIG& at(size_t pos) = 0;
 
     virtual void load(const ProjSIG* data, UINT32 num) = 0;
 };
